--- a/src/applications/pensions/tests/e2e/helpers/pageHooks.js
+++ b/src/applications/pensions/tests/e2e/helpers/pageHooks.js
@@ -1,15 +1,3 @@
-import testForm from 'platform/testing/e2e/cypress/support/form-tester';
-import { createTestConfig } from 'platform/testing/e2e/cypress/support/form-tester/utilities';
-
-import kitchenSinkFixture from 'vets-json-schema/dist/21P-527EZ-KITCHEN_SINK-cypress-example.json';
-import overflowFixture from 'vets-json-schema/dist/21P-527EZ-OVERFLOW-cypress-example.json';
-import simpleFixture from 'vets-json-schema/dist/21P-527EZ-SIMPLE-cypress-example.json';
-
-import mockUser from '../fixtures/mocks/user.json';
-import formConfig from '../../config/form';
-import manifest from '../../manifest.json';
-import setupCypress, { cypressBeforeAllSetup } from './cypress.setup';
-
 import {
   fillAddressWebComponentPattern,
   fillCareExpensesPage,
@@ -24,9 +12,9 @@ import {
   fillSpouseMarriagesPage,
   fillVaMedicalCentersPage,
   shouldNotHaveValidationErrors,
-} from './helpers';
+} from './index';
 
-import pagePaths from './pagePaths';
+import pagePaths from '../pagePaths';
 
 const replaceDefaultPostHook = ({ afterHook }) => {
   afterHook(() => {
@@ -40,7 +28,7 @@ const replaceDefaultBehavior = context => {
   replaceDefaultPostHook(context);
 };
 
-export const pageHooks = cy => ({
+const pageHooks = {
   introduction: () => {
     // skip wizard
     cy.findAllByText(/start the pension application/i)
@@ -227,32 +215,6 @@ export const pageHooks = cy => ({
       });
     });
   },
-});
+};
 
-const testConfig = createTestConfig(
-  {
-    useWebComponentFields: true,
-    appName: 'Pensions',
-    dataPrefix: 'data',
-    dataDir: null,
-    dataSets: [
-      { title: 'kitchen-sink', data: kitchenSinkFixture },
-      { title: 'overflow', data: overflowFixture },
-      { title: 'simple', data: simpleFixture },
-    ],
-    pageHooks: pageHooks(cy),
-    setup: () => {
-      cypressBeforeAllSetup();
-    },
-    setupPerTest: () => {
-      cy.login(mockUser);
-      setupCypress(cy);
-    },
-
-    // skip: [],
-  },
-  manifest,
-  formConfig,
-);
-
-testForm(testConfig);
+export default pageHooks;