… src/applications/pensions/tests/e2e/pensions-keyboard-only.cypress.spec.js