Are you removing, renaming or moving a folder in this PR?
- [x] No, I'm not changing any folders (skip to TeamSites and delete the rest of this section)
- [ ] Yes, I'm removing, renaming or moving a folder

 Did you change site-wide styles, platform utilities or other infrastructure?
- [x] No
- [ ] Yes

## Summary
It's normal for Cypress tests to take **2-3x longer** in CI/CD due to resource limitations, network latency, and logging overhead. To improve the efficiency of our CI/CD pipeline, the end-to-end (E2E) tests for the Pension Benefits form need to be updated to support concurrent execution. Currently, multiple test suites are grouped into a single file, which prevents them from being run concurrently. This task involves separating each test suite into its own file, allowing the Platform team to configure the tests to run concurrently within the pipeline. Additionally, the overflow test suite has been disabled in CI to expedite the process. The next step is to notify the Platform team that tests are ready for concurrent execution in the CI/CD pipeline. 

## Issue
- https://github.com/department-of-veterans-affairs/va.gov-team/issues/97619

## Related issue(s)
n/a

## Associated Pull Request(s)
n/a

## How to run in local environment
1. Check out this branch locally
2. Run `vets-website` and `vets-api`

## How to verify
### Run each test individually
#### Simple
`yarn cy:run --spec src/applications/pensions/tests/e2e/pensions-simple.cypress.spec.js`
`  Spec                                              Tests  Passing  Failing  Pending  Skipped  
  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
  │ ✔  pensions-simple.cypress.spec.js          00:18        1        1        -        -        - │
  └────────────────────────────────────────────────────────────────────────────────────────────────┘
    ✔  All specs passed!                        00:18        1        1        -        -        -  `

#### Kitchen Sink
`yarn cy:run --spec src/applications/pensions/tests/e2e/pensions-kitchen-sink.cypress.spec.js`
` Spec                                              Tests  Passing  Failing  Pending  Skipped  
  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
  │ ✔  pensions-kitchen-sink.cypress.spec.      01:26        1        1        -        -        - │
  │    js                                                                                          │
  └────────────────────────────────────────────────────────────────────────────────────────────────┘
    ✔  All specs passed!                        01:26        1        1        -        -        -  `

#### Overflow
`yarn cy:run --spec src/applications/pensions/tests/e2e/pensions-overflow.cypress.spec.js`
`     Spec                                              Tests  Passing  Failing  Pending  Skipped  
  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
  │ ✔  pensions-overflow.cypress.spec.js        01:46        1        1        -        -        - │
  └────────────────────────────────────────────────────────────────────────────────────────────────┘
    ✔  All specs passed!                        01:46        1        1        -        -        -  `

#### Keyboard Only
`yarn cy:run --spec src/applications/pensions/tests/e2e/pensions-keyboard-only.cypress.spec.js`
`  Spec                                              Tests  Passing  Failing  Pending  Skipped  
  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
  │ ✔  pensions-keyboard-only.cypress.spec      00:41        2        2        -        -        - │
  │    .js                                                                                         │
  └────────────────────────────────────────────────────────────────────────────────────────────────┘
    ✔  All specs passed!                        00:41        2        2        -        -   `


## What areas of the site does it impact?
Pension Benefits Application

## Screenshots
### CI/CD
#### E2E Test Stability Review Runs
![Screenshot 2025-01-29 at 9 10 48 AM](https://github.com/user-attachments/assets/8dfc1266-8320-4a98-97a8-d1bbeb386a48)

![Screenshot 2025-01-29 at 10 13 05 AM](https://github.com/user-attachments/assets/01f9c95a-69f7-43fe-b962-13e4b768c284)

### Mobile (if applicable)
| Before      | After       |
| ----------- | ----------- |
| Upload file | Upload file |

### Quality Assurance & Testing
- [ ] New unit tests (if applicable)
- [x] New E2E tests added (if applicable)
- [x] Existing unit tests and integration tests are passing
- [x] Existing E2E tests are passing
- [x] No sensitive information (i.e. PII/credentials/internal URLs/etc.) is captured in logging, hardcoded, or specs
- [ ] Linting warnings have been addressed
- [ ] Documentation has been updated ([link to documentation](#) \*if necessary)
- [ ] Screenshot style, layout and content matches the design references
- [ ] [Accessibility testing](https://depo-platform-documentation.scrollhelp.site/developer-docs/wcag-2-1-success-criteria-and-foundational-testing) has been performed

### Error Handling
- [x] Browser console contains no warnings or errors.
- [ ] Events are being sent to the appropriate logging solution
- [ ] Feature/bug has a monitor built into Datadog or Grafana (if applicable)

### Authentication
- [x] Did you login to a local build and verify all authenticated routes work as expected with a test user

## Requested Feedback
(OPTIONAL) _What should the reviewers know in addition to the above. Is there anything specific you wish the reviewer to assist with. Do you have any concerns with this PR, why?_