@ToddWebDev 2025-01-27
Subsequent CI/CD runs:
| Test Suite | Time |
|--------|--------|
| Simple | 43s |
| Kitchen Sink | 5m57s |
| Keyboard Only | 6m 42s |