@ToddWebDev 2025-01-27
Subsequent CI/CD runs:
| Test Suite | Time |
|--------|--------|
| Simple | 41s |
| Kitchen Sink | 6m 9s |
| Keyboard Only | 7m 1s |