@ToddWebDev 2025-01-28
Subsequent CI/CD runs:
```Spec                                              Tests  Passing  Failing  Pending  Skipped  
  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
  │ ✔  pensions-kitchen-sink.cypress.spec.      08:12        1        1        -        -        - │
  │    js                                                                                          │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-keyboard-only.cypress.spec      01:37        4        2        -        2        - │
  │    .js                                                                                         │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-simple.cypress.spec.js          00:54        1        1        -        -        - │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-overflow.cypress.spec.js         93ms        1        -        -        1        - │
  └────────────────────────────────────────────────────────────────────────────────────────────────┘
    ✔  All specs passed!                        10:44        7        4        -        3        -