@ToddWebDev 2025-01-28
``` Spec                                              Tests  Passing  Failing  Pending  Skipped  
  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
  │ ✔  pensions-kitchen-sink.cypress.spec.      07:41        1        1        -        -        - │
  │    js                                                                                          │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-keyboard-only.cypress.spec      00:43        4        2        -        2        - │
  │    .js                                                                                         │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-simple.cypress.spec.js          00:52        1        1        -        -        - │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-overflow.cypress.spec.js         94ms        1        -        -        1        - │
  └────────────────────────────────────────────────────────────────────────────────────────────────┘
    ✔  All specs passed!                        09:18        7        4        -        3        -  ```