@ToddWebDev 2025-01-29
```     Spec                                              Tests  Passing  Failing  Pending  Skipped  
  ┌────────────────────────────────────────────────────────────────────────────────────────────────┐
  │ ✔  pensions-keyboard-only.cypress.spec      01:40        4        2        -        2        - │
  │    .js                                                                                         │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-simple.cypress.spec.js          00:52        1        1        -        -        - │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-kitchen-sink.cypress.spec.      08:07        1        1        -        -        - │
  │    js                                                                                          │
  ├────────────────────────────────────────────────────────────────────────────────────────────────┤
  │ ✔  pensions-overflow.cypress.spec.js         92ms        1        -        -        1        - │
  └────────────────────────────────────────────────────────────────────────────────────────────────┘
    ✔  All specs passed!                        10:41        7        4        -        3        -  ```