@ToddWebDev 2025-01-30
After discussion with the Frontend CoP: 
Platform doesn't subscribe to Cypress Dashboard native parallelization, but individual test suites will be run as separate github actions that utilize parallel processes to accomplish this in an unconventional way.